Injects component below Amazon price (keeps old injectRating.js for reference)
