+++ src/js/injectRatingOld.js
@@ -1,0 +1,131 @@
+const greenColor = "#2a6355";
+
+window.addEventListener("load", (e)=>{
+    //INJECTION CODE
+    let sustainabilityScore = 4;
+    let cost = 3;
+    let name = "Peanut Butter";
+    let carbonAmount = '1';
+
+    addRatingButton(sustainabilityScore, name, carbonAmount, cost);
+
+
+});
+function addRatingButton(sustainabilityScore, name, carbonAmount, cost) {
+    const amazonPriceSection = document.getElementById("unifiedPrice_feature_div");
+    let ratingNode = document.createElement("DIV");
+    ratingNode.style.padding = "5px";
+    ratingNode.style.borderRadius = "15px";
+    ratingNode.style.alignContent = "center";
+    ratingNode.style.display = "flex";
+    ratingNode.style.alignItems = "center";
+    ratingNode.style.width = "160px";
+    ratingNode.style.height = "35px";
+    ratingNode.style.whiteSpace = "nowrap";
+    let mapIcon = document.createElement("mapIcon");
+    mapIcon.innerHTML = miniWhiteMapSVG;
+    mapIcon.style.paddingRight = "10px";
+    let buttonRatingText = document.createElement("rating");
+    buttonRatingText.style.borderStyle = "solid";
+    buttonRatingText.style.borderWidth = "thin";
+    buttonRatingText.style.padding = "2px";
+    buttonRatingText.style.marginRight = "10px";
+    buttonRatingText.innerText = sustainabilityScore.toString();
+    let climateText = document.createElement("climateText");
+    climateText.innerText = "Climate Rating >";
+    climateText.style.fontSize = "12px";
+    climateText.style.fontWeight = "600";
+    ratingNode.appendChild(mapIcon);
+    ratingNode.appendChild(buttonRatingText);
+    ratingNode.appendChild(climateText);
+    ratingNode.style.color = "White";
+    ratingNode.style.background = greenColor;
+    amazonPriceSection.prepend(ratingNode);
+    //MODAL CODE
+    const modalGlobe = document.createElement("mapIcon");
+    modalGlobe.innerHTML = bigGreenMapSVG;
+    let productName = document.createElement("h1");
+    productName.style.color = greenColor;
+    productName.innerText = name.toUpperCase();
+    productName.style.fontSize = "14px";
+    let modalPopup = document.createElement("climatePopup");
+    modalPopup.style.display = "none";
+    modalPopup.style.color = 'black';
+    modalPopup.style.backgroundColor = "white";
+    modalPopup.style.position = "relative";
+    modalPopup.style.left = "7px";
+    modalPopup.style.border = "solid";
+    modalPopup.style.borderWidth = "thin";
+    modalPopup.style.borderColor = "#e8ebe8";
+    modalPopup.style.borderRadius = "20px";
+    modalPopup.style.padding = "20px";
+    modalPopup.style.flexDirection = "column";
+    modalPopup.style.width = "200px";
+    modalPopup.style.zIndex = "100";
+    modalPopup.appendChild(modalGlobe);
+    modalPopup.appendChild(productName);
+    modalPopup.appendChild(makeDivider());
+    modalPopup.appendChild(makeTitleWidget("Carbon"));
+    modalPopup.appendChild(makeDataWidget(carbonAmount, "gCO2e"));
+    modalPopup.appendChild(makeDivider());
+    modalPopup.appendChild(makeTitleWidget("Cost"));
+    modalPopup.appendChild(makeDataWidget(cost, "US$"));
+    modalPopup.appendChild(makeDivider());
+    modalPopup.appendChild(makeTitleWidget("Sustainability"));
+    modalPopup.appendChild(makeDataWidget(sustainabilityScore, ""));
+    ratingNode.appendChild(modalPopup);
+    ratingNode.addEventListener("mouseover", (event) => {
+        console.log("Moused over climate rating");
+        showClimateInfoPopup(modalPopup);
+    });
+    ratingNode.addEventListener("mouseout", (event) => {
+        console.log("Moused away from climate rating");
+        closeClimateInfoPopup(modalPopup);
+    });
+}
+
+function makeDivider(){
+    let divider = document.createElement("div");
+    divider.style.height = "1px";
+    divider.style.marginBottom = "15px";
+    divider.style.marginTop = "15px";
+
+    divider.style.backgroundColor = "lightgray";
+    return divider;
+}
+function makeTitleWidget(title){
+    let titleWidget = document.createElement("h1");
+    titleWidget.style.fontSize = "16px";
+    titleWidget.style.color = "black";
+    titleWidget.style.fontWeight = "300";
+    titleWidget.innerText = title.toUpperCase();
+    return titleWidget;
+}
+function makeDataWidget(amount, measurement){
+    let dataWidget = document.createElement("measurementContainer");
+    dataWidget.style.color = greenColor;
+
+    let amountWidget = document.createElement("amount");
+    amountWidget.style.fontSize = "28px";
+    amountWidget.innerText = amount.toString();
+
+    let measurementWidget = document.createElement("measurement");
+    measurementWidget.style.fontSize = "13px";
+    measurementWidget.innerText = measurement;
+
+    dataWidget.appendChild(amountWidget);
+    dataWidget.appendChild(measurementWidget);
+    
+    return dataWidget;
+}
+function closeClimateInfoPopup(modalPopup){
+    modalPopup.style.display = "none";
+}
+function showClimateInfoPopup(modalPopup){
+    modalPopup.style.display = "block";
+}
+
+
+const bigGreenMapSVG = '<svg height="45px" viewBox="-344.097 -1906.932 20.163 20.163"><path fill="'+greenColor+'" id="Path_4667" d="M -323.9360046386719 -1896.954956054688 C -323.9389953613281 -1897.25 -323.9540100097656 -1897.54296875 -323.9809875488281 -1897.8310546875 C -324.0750122070312 -1898.7919921875 -324.3039855957031 -1899.712036132812 -324.6489868164062 -1900.576049804688 C -324.6740112304688 -1900.640014648438 -324.697998046875 -1900.704956054688 -324.7250061035156 -1900.76904296875 C -325.1839904785156 -1901.85205078125 -325.8280029296875 -1902.838989257812 -326.6159973144531 -1903.68994140625 C -326.6679992675781 -1903.745971679688 -326.7210083007812 -1903.800048828125 -326.7730102539062 -1903.85400390625 C -326.9219970703125 -1904.009033203125 -327.0750122070312 -1904.160034179688 -327.2340087890625 -1904.305053710938 C -329.0260009765625 -1905.93603515625 -331.406005859375 -1906.932006835938 -334.0150146484375 -1906.932006835938 C -336.64599609375 -1906.932006835938 -339.0429992675781 -1905.91796875 -340.8410034179688 -1904.261962890625 C -341.2590026855469 -1903.8759765625 -341.6440124511719 -1903.453979492188 -341.9930114746094 -1903.003051757812 C -343.3099975585938 -1901.300048828125 -344.0969848632812 -1899.166015625 -344.0969848632812 -1896.850952148438 C -344.0969848632812 -1891.291015625 -339.5740051269531 -1886.76904296875 -334.0150146484375 -1886.76904296875 C -330.10400390625 -1886.76904296875 -326.7090148925781 -1889.008056640625 -325.0379943847656 -1892.27099609375 C -324.6809997558594 -1892.968017578125 -324.4030151367188 -1893.712036132812 -324.2149963378906 -1894.490966796875 C -324.1679992675781 -1894.68896484375 -324.1260070800781 -1894.886962890625 -324.0910034179688 -1895.088012695312 C -323.989013671875 -1895.661010742188 -323.9339904785156 -1896.25 -323.9339904785156 -1896.850952148438 C -323.9330139160156 -1896.885986328125 -323.9349975585938 -1896.920043945312 -323.9360046386719 -1896.954956054688 Z M -325.197998046875 -1901.305053710938 L -325.135009765625 -1901.3759765625 C -325.052001953125 -1901.215942382812 -324.9750061035156 -1901.052001953125 -324.9010009765625 -1900.886962890625 L -325.0039978027344 -1900.890991210938 L -325.197998046875 -1900.864013671875 L -325.197998046875 -1901.305053710938 Z M -326.7680053710938 -1903.201049804688 L -326.7659912109375 -1903.68603515625 C -326.5950012207031 -1903.505004882812 -326.4309997558594 -1903.318969726562 -326.2730102539062 -1903.1259765625 L -326.468994140625 -1902.8349609375 L -327.1530151367188 -1902.840942382812 L -327.1960144042969 -1902.983032226562 L -326.7680053710938 -1903.201049804688 Z M -339.8739929199219 -1904.406982421875 L -339.8739929199219 -1904.426025390625 L -339.656005859375 -1904.426025390625 L -339.6369934082031 -1904.5 L -339.281005859375 -1904.5 L -339.281005859375 -1904.344970703125 L -339.3840026855469 -1904.2080078125 L -339.8739929199219 -1904.2080078125 L -339.8739929199219 -1904.406982421875 L -339.8739929199219 -1904.406982421875 Z M -339.5260009765625 -1903.921997070312 C -339.5260009765625 -1903.921997070312 -339.3080139160156 -1903.958984375 -339.2900085449219 -1903.958984375 C -339.27099609375 -1903.958984375 -339.2900085449219 -1903.741943359375 -339.2900085449219 -1903.741943359375 L -339.7799987792969 -1903.711059570312 L -339.8739929199219 -1903.822998046875 L -339.5260009765625 -1903.921997070312 Z M -324.5050048828125 -1899.609008789062 L -325.2999877929688 -1899.609008789062 L -325.7850036621094 -1899.969970703125 L -326.2950134277344 -1899.921020507812 L -326.2950134277344 -1899.609008789062 L -326.4559936523438 -1899.609008789062 L -326.6300048828125 -1899.734008789062 L -327.5130004882812 -1899.9580078125 L -327.5130004882812 -1900.530029296875 L -328.6319885253906 -1900.442016601562 L -328.97900390625 -1900.255981445312 L -329.4230041503906 -1900.255981445312 L -329.6409912109375 -1900.277954101562 L -330.1799926757812 -1899.97900390625 L -330.1799926757812 -1899.4150390625 L -331.2829895019531 -1898.619995117188 L -331.1910095214844 -1898.280029296875 L -330.9679870605469 -1898.280029296875 L -331.0260009765625 -1897.9560546875 L -331.1839904785156 -1897.899047851562 L -331.1919860839844 -1897.052978515625 L -330.239013671875 -1895.968017578125 L -329.8250122070312 -1895.968017578125 L -329.7999877929688 -1896.034057617188 L -329.0539855957031 -1896.034057617188 L -328.8389892578125 -1896.233032226562 L -328.4169921875 -1896.233032226562 L -328.1849975585938 -1896.0009765625 L -327.5539855957031 -1895.93603515625 L -327.6380004882812 -1895.098022460938 L -326.9389953613281 -1893.863037109375 L -327.3070068359375 -1893.159057617188 L -327.2820129394531 -1892.828002929688 L -326.9920043945312 -1892.537963867188 L -326.9920043945312 -1891.740966796875 L -326.6109924316406 -1891.22900390625 L -326.6109924316406 -1890.567016601562 L -326.2820129394531 -1890.567016601562 C -328.1130065917969 -1888.318969726562 -330.8989868164062 -1886.880004882812 -334.0159912109375 -1886.880004882812 C -339.5130004882812 -1886.879028320312 -343.9859924316406 -1891.35302734375 -343.9859924316406 -1896.850952148438 C -343.9859924316406 -1898.234985351562 -343.7019958496094 -1899.553955078125 -343.1900024414062 -1900.753051757812 L -343.1900024414062 -1901.06396484375 L -342.8340148925781 -1901.496948242188 C -342.7099914550781 -1901.73095703125 -342.5769958496094 -1901.958984375 -342.4349975585938 -1902.181030273438 L -342.4190063476562 -1902 L -342.8330078125 -1901.496948242188 C -342.9609985351562 -1901.255004882812 -343.0809936523438 -1901.006958007812 -343.1900024414062 -1900.753051757812 L -343.1900024414062 -1900.18603515625 L -342.7749938964844 -1899.985961914062 L -342.7749938964844 -1899.197021484375 L -342.3779907226562 -1898.52001953125 L -342.0549926757812 -1898.469970703125 L -342.0130004882812 -1898.702026367188 L -342.3949890136719 -1899.2900390625 L -342.468994140625 -1899.862060546875 L -342.2460021972656 -1899.862060546875 L -342.1520080566406 -1899.274047851562 L -341.6010131835938 -1898.469970703125 L -341.7430114746094 -1898.2099609375 L -341.3919982910156 -1897.673950195312 L -340.5220031738281 -1897.458984375 L -340.5220031738281 -1897.599975585938 L -340.1740112304688 -1897.550048828125 L -340.2070007324219 -1897.302001953125 L -339.9339904785156 -1897.251953125 L -339.510986328125 -1897.135986328125 L -338.9150085449219 -1896.45703125 L -338.1530151367188 -1896.399047851562 L -338.0780029296875 -1895.776977539062 L -338.6000061035156 -1895.412963867188 L -338.6239929199219 -1894.858032226562 L -338.6990051269531 -1894.517944335938 L -337.9450073242188 -1893.573974609375 L -337.8869934082031 -1893.25 C -337.8869934082031 -1893.25 -337.6130065917969 -1893.176025390625 -337.5799865722656 -1893.176025390625 C -337.5469970703125 -1893.176025390625 -336.9670104980469 -1892.737060546875 -336.9670104980469 -1892.737060546875 L -336.9670104980469 -1891.030029296875 L -336.760009765625 -1890.972045898438 L -336.9010009765625 -1890.18505859375 L -336.5530090332031 -1889.719970703125 L -336.6170043945312 -1888.93994140625 L -336.1570129394531 -1888.131958007812 L -335.5669860839844 -1887.615966796875 L -334.9729919433594 -1887.60498046875 L -334.9150085449219 -1887.796020507812 L -335.3510131835938 -1888.162963867188 L -335.3259887695312 -1888.345947265625 L -335.2479858398438 -1888.568969726562 L -335.2319946289062 -1888.796997070312 L -335.5260009765625 -1888.806030273438 L -335.6749877929688 -1888.993041992188 L -335.4299926757812 -1889.22802734375 L -335.3970031738281 -1889.406005859375 L -335.6709899902344 -1889.484008789062 L -335.6549987792969 -1889.650024414062 L -335.2650146484375 -1889.708984375 L -334.6719970703125 -1889.993041992188 L -334.4739990234375 -1890.358032226562 L -333.8519897460938 -1891.154052734375 L -333.9930114746094 -1891.776000976562 L -333.802001953125 -1892.107055664062 L -333.2309875488281 -1892.088989257812 L -332.8460083007812 -1892.39404296875 L -332.7219848632812 -1893.593994140625 L -332.2950134277344 -1894.135986328125 L -332.2200012207031 -1894.484985351562 L -332.6099853515625 -1894.609008789062 L -332.8670043945312 -1895.031005859375 L -333.7449951171875 -1895.0400390625 L -334.4410095214844 -1895.305053710938 L -334.4739990234375 -1895.802001953125 L -334.7059936523438 -1896.2080078125 L -335.3359985351562 -1896.217041015625 L -335.7000122070312 -1896.787963867188 L -336.0230102539062 -1896.946044921875 L -336.0400085449219 -1896.77197265625 L -336.6279907226562 -1896.737060546875 L -336.8429870605469 -1897.036987304688 L -337.4570007324219 -1897.161010742188 L -337.9620056152344 -1896.577026367188 L -338.7569885253906 -1896.712036132812 L -338.8150024414062 -1897.609008789062 L -339.39599609375 -1897.7080078125 L -339.1629943847656 -1898.14794921875 L -339.22900390625 -1898.400024414062 L -339.9920043945312 -1897.890014648438 L -340.4719848632812 -1897.948974609375 L -340.6430053710938 -1898.323974609375 L -340.5390014648438 -1898.7099609375 L -340.2739868164062 -1899.197998046875 L -339.6650085449219 -1899.505981445312 L -338.489013671875 -1899.506958007812 L -338.4920043945312 -1899.14794921875 L -338.0690002441406 -1898.951049804688 L -338.1029968261719 -1899.56396484375 L -337.7980041503906 -1899.870971679688 L -337.1830139160156 -1900.275024414062 L -337.1409912109375 -1900.558959960938 L -336.5280151367188 -1901.197021484375 L -335.8760070800781 -1901.557983398438 L -335.9339904785156 -1901.60595703125 L -335.4920043945312 -1902.02099609375 L -335.3299865722656 -1901.97802734375 L -335.2560119628906 -1901.885009765625 L -335.0889892578125 -1902.071044921875 L -335.0469970703125 -1902.089965820312 L -335.2309875488281 -1902.115966796875 L -335.4169921875 -1902.177978515625 L -335.4169921875 -1902.35595703125 L -335.3190002441406 -1902.43701171875 L -335.1010131835938 -1902.43701171875 L -335.0010070800781 -1902.39404296875 L -334.9150085449219 -1902.218994140625 L -334.8099975585938 -1902.234985351562 L -334.8099975585938 -1902.2509765625 L -334.7789916992188 -1902.239990234375 L -334.4739990234375 -1902.286987304688 L -334.4299926757812 -1902.43603515625 L -334.2560119628906 -1902.392944335938 L -334.2560119628906 -1902.23095703125 L -334.4179992675781 -1902.119995117188 L -334.4169921875 -1902.119995117188 L -334.3940124511719 -1901.942016601562 L -333.8399963378906 -1901.77197265625 L -333.8389892578125 -1901.765014648438 L -333.7109985351562 -1901.776000976562 L -333.7030029296875 -1902.015991210938 L -334.1419982910156 -1902.214965820312 L -334.1669921875 -1902.3310546875 L -333.802001953125 -1902.454956054688 L -333.7860107421875 -1902.802978515625 L -334.1669921875 -1903.035034179688 L -334.1919860839844 -1903.623046875 L -334.7139892578125 -1903.365966796875 L -334.9039916992188 -1903.365966796875 L -334.85400390625 -1903.81396484375 L -335.5650024414062 -1903.98095703125 L -335.8590087890625 -1903.759033203125 L -335.8590087890625 -1903.08203125 L -336.3880004882812 -1902.913940429688 L -336.5989990234375 -1902.473022460938 L -336.8290100097656 -1902.43603515625 L -336.8290100097656 -1903 L -337.3259887695312 -1903.068969726562 L -337.5750122070312 -1903.23095703125 L -337.6749877929688 -1903.597045898438 L -336.7860107421875 -1904.116943359375 L -336.3510131835938 -1904.25 L -336.3070068359375 -1903.95703125 L -336.0650024414062 -1903.969970703125 L -336.0459899902344 -1904.116943359375 L -335.7929992675781 -1904.152954101562 L -335.7879943847656 -1904.203979492188 L -335.8970031738281 -1904.25 L -335.9219970703125 -1904.405029296875 L -335.6109924316406 -1904.431030273438 L -335.4230041503906 -1904.626953125 L -335.4129943847656 -1904.640991210938 L -335.4100036621094 -1904.640991210938 L -335.3529968261719 -1904.699951171875 L -334.697998046875 -1904.781982421875 L -334.4089965820312 -1904.536010742188 L -335.1679992675781 -1904.131958007812 L -334.2019958496094 -1903.904052734375 L -334.0769958496094 -1904.22705078125 L -333.6539916992188 -1904.22705078125 L -333.5050048828125 -1904.509033203125 L -333.8039855957031 -1904.5830078125 L -333.8039855957031 -1904.93896484375 L -334.739990234375 -1905.35400390625 L -335.385986328125 -1905.280029296875 L -335.7510070800781 -1905.088989257812 L -335.7260131835938 -1904.625 L -336.1069946289062 -1904.682983398438 L -336.1659851074219 -1904.93896484375 L -335.7999877929688 -1905.27099609375 L -336.4630126953125 -1905.303955078125 L -336.6539916992188 -1905.246948242188 L -336.7359924316406 -1905.02294921875 L -336.4880065917969 -1904.98095703125 L -336.5369873046875 -1904.733032226562 L -336.9599914550781 -1904.7080078125 L -337.0260009765625 -1904.5419921875 L -337.6380004882812 -1904.525024414062 C -337.6380004882812 -1904.525024414062 -337.6549987792969 -1904.873046875 -337.6799926757812 -1904.873046875 C -337.7049865722656 -1904.873046875 -337.2000122070312 -1904.880981445312 -337.2000122070312 -1904.880981445312 L -336.8349914550781 -1905.238037109375 L -337.0339965820312 -1905.338012695312 L -337.2990112304688 -1905.079956054688 L -337.739013671875 -1905.10498046875 L -338.0039978027344 -1905.469970703125 L -338.5669860839844 -1905.469970703125 L -339.1549987792969 -1905.031005859375 L -338.6159973144531 -1905.031005859375 L -338.5669860839844 -1904.873046875 L -338.7070007324219 -1904.743041992188 L -338.1099853515625 -1904.725952148438 L -338.0190124511719 -1904.510986328125 L -338.6910095214844 -1904.536010742188 L -338.7229919433594 -1904.702026367188 L -339.14599609375 -1904.79296875 L -339.3699951171875 -1904.9169921875 L -339.8720092773438 -1904.912963867188 C -338.2269897460938 -1906.112060546875 -336.2030029296875 -1906.822021484375 -334.0150146484375 -1906.822021484375 C -331.4939880371094 -1906.822021484375 -329.1900024414062 -1905.879028320312 -327.4320068359375 -1904.3310546875 L -327.5509948730469 -1904.119018554688 L -328.010986328125 -1903.93896484375 L -328.2049865722656 -1903.72802734375 L -328.1600036621094 -1903.483032226562 L -327.9230041503906 -1903.449951171875 L -327.7799987792969 -1903.093017578125 L -327.3699951171875 -1903.256958007812 L -327.3030090332031 -1902.779052734375 L -327.4259948730469 -1902.779052734375 L -327.7619934082031 -1902.828979492188 L -328.135009765625 -1902.765991210938 L -328.4960021972656 -1902.256958007812 L -329.010986328125 -1902.176025390625 L -329.0859985351562 -1901.734008789062 L -328.8680114746094 -1901.682983398438 L -328.9309997558594 -1901.399047851562 L -329.4429931640625 -1901.501953125 L -329.9129943847656 -1901.399047851562 L -330.0130004882812 -1901.137939453125 L -329.9309997558594 -1900.588989257812 L -329.656005859375 -1900.4599609375 L -329.1929931640625 -1900.463012695312 L -328.8810119628906 -1900.490966796875 L -328.7860107421875 -1900.739013671875 L -328.2980041503906 -1901.373046875 L -327.9760131835938 -1901.307983398438 L -327.6600036621094 -1901.593994140625 L -327.6010131835938 -1901.369995117188 L -326.8219909667969 -1900.844970703125 L -326.9179992675781 -1900.717041015625 L -327.2680053710938 -1900.735961914062 L -327.1340026855469 -1900.545043945312 L -326.9179992675781 -1900.496948242188 L -326.6650085449219 -1900.60302734375 L -326.6700134277344 -1900.907958984375 L -326.5580139160156 -1900.963989257812 L -326.6480102539062 -1901.06005859375 L -327.1669921875 -1901.348999023438 L -327.3039855957031 -1901.734008789062 L -326.8729858398438 -1901.734008789062 L -326.7340087890625 -1901.597045898438 L -326.3630065917969 -1901.276977539062 L -326.3469848632812 -1900.889038085938 L -325.9620056152344 -1900.47900390625 L -325.8179931640625 -1901.041015625 L -325.552001953125 -1901.18701171875 L -325.5010070800781 -1900.72705078125 L -325.2409973144531 -1900.441040039062 L -324.7210083007812 -1900.449951171875 C -324.6210021972656 -1900.191040039062 -324.5299987792969 -1899.927978515625 -324.4509887695312 -1899.660034179688 L -324.5050048828125 -1899.609008789062 Z M -338.9540100097656 -1902.779052734375 L -338.6929931640625 -1902.902954101562 L -338.4570007324219 -1902.847045898438 L -338.5379943847656 -1902.530029296875 L -338.7929992675781 -1902.448974609375 L -338.9540100097656 -1902.779052734375 Z M -337.5690002441406 -1902.032958984375 L -337.5690002441406 -1901.828002929688 L -338.1650085449219 -1901.828002929688 L -338.3880004882812 -1901.890014648438 L -338.3320007324219 -1902.032958984375 L -338.0459899902344 -1902.151977539062 L -337.6549987792969 -1902.151977539062 L -337.6549987792969 -1902.032958984375 L -337.5690002441406 -1902.032958984375 Z M -337.2940063476562 -1901.746948242188 L -337.2940063476562 -1901.548950195312 L -337.4440002441406 -1901.452026367188 L -337.6300048828125 -1901.41796875 L -337.6300048828125 -1901.746948242188 L -337.2940063476562 -1901.746948242188 Z M -337.4620056152344 -1901.828002929688 L -337.4620056152344 -1902.06494140625 L -337.2569885253906 -1901.878051757812 L -337.4620056152344 -1901.828002929688 Z M -337.3689880371094 -1901.349975585938 L -337.3689880371094 -1901.156982421875 L -337.5119934082031 -1901.014038085938 L -337.8280029296875 -1901.014038085938 L -337.7789916992188 -1901.23095703125 L -337.6289978027344 -1901.244018554688 L -337.5979919433594 -1901.317993164062 L -337.3689880371094 -1901.349975585938 Z M -338.1579895019531 -1901.746948242188 L -337.8290100097656 -1901.746948242188 L -338.2510070800781 -1901.156982421875 L -338.4249877929688 -1901.25 L -338.3880004882812 -1901.4990234375 L -338.1579895019531 -1901.746948242188 Z M -336.8089904785156 -1901.41796875 L -336.8089904785156 -1901.224975585938 L -337.1260070800781 -1901.224975585938 L -337.2130126953125 -1901.349975585938 L -337.2130126953125 -1901.530029296875 L -337.18798828125 -1901.530029296875 L -336.8089904785156 -1901.41796875 Z M -337.1019897460938 -1901.682983398438 L -337.0119934082031 -1901.777954101562 L -336.8599853515625 -1901.682983398438 L -336.9819946289062 -1901.5830078125 L -337.1019897460938 -1901.682983398438 Z M -324.343994140625 -1899.10595703125 L -324.31201171875 -1899.142944335938 C -324.2990112304688 -1899.087036132812 -324.2860107421875 -1899.030029296875 -324.2730102539062 -1898.973022460938 L -324.343994140625 -1899.10595703125 Z"></path></svg>'
+
+const miniWhiteMapSVG = '<svg height="20px" viewBox="-344.097 -1906.932 20.163 20.163"><path fill="rgba(255,255,255,1)" id="Path_4667" d="M -323.9360046386719 -1896.954956054688 C -323.9389953613281 -1897.25 -323.9540100097656 -1897.54296875 -323.9809875488281 -1897.8310546875 C -324.0750122070312 -1898.7919921875 -324.3039855957031 -1899.712036132812 -324.6489868164062 -1900.576049804688 C -324.6740112304688 -1900.640014648438 -324.697998046875 -1900.704956054688 -324.7250061035156 -1900.76904296875 C -325.1839904785156 -1901.85205078125 -325.8280029296875 -1902.838989257812 -326.6159973144531 -1903.68994140625 C -326.6679992675781 -1903.745971679688 -326.7210083007812 -1903.800048828125 -326.7730102539062 -1903.85400390625 C -326.9219970703125 -1904.009033203125 -327.0750122070312 -1904.160034179688 -327.2340087890625 -1904.305053710938 C -329.0260009765625 -1905.93603515625 -331.406005859375 -1906.932006835938 -334.0150146484375 -1906.932006835938 C -336.64599609375 -1906.932006835938 -339.0429992675781 -1905.91796875 -340.8410034179688 -1904.261962890625 C -341.2590026855469 -1903.8759765625 -341.6440124511719 -1903.453979492188 -341.9930114746094 -1903.003051757812 C -343.3099975585938 -1901.300048828125 -344.0969848632812 -1899.166015625 -344.0969848632812 -1896.850952148438 C -344.0969848632812 -1891.291015625 -339.5740051269531 -1886.76904296875 -334.0150146484375 -1886.76904296875 C -330.10400390625 -1886.76904296875 -326.7090148925781 -1889.008056640625 -325.0379943847656 -1892.27099609375 C -324.6809997558594 -1892.968017578125 -324.4030151367188 -1893.712036132812 -324.2149963378906 -1894.490966796875 C -324.1679992675781 -1894.68896484375 -324.1260070800781 -1894.886962890625 -324.0910034179688 -1895.088012695312 C -323.989013671875 -1895.661010742188 -323.9339904785156 -1896.25 -323.9339904785156 -1896.850952148438 C -323.9330139160156 -1896.885986328125 -323.9349975585938 -1896.920043945312 -323.9360046386719 -1896.954956054688 Z M -325.197998046875 -1901.305053710938 L -325.135009765625 -1901.3759765625 C -325.052001953125 -1901.215942382812 -324.9750061035156 -1901.052001953125 -324.9010009765625 -1900.886962890625 L -325.0039978027344 -1900.890991210938 L -325.197998046875 -1900.864013671875 L -325.197998046875 -1901.305053710938 Z M -326.7680053710938 -1903.201049804688 L -326.7659912109375 -1903.68603515625 C -326.5950012207031 -1903.505004882812 -326.4309997558594 -1903.318969726562 -326.2730102539062 -1903.1259765625 L -326.468994140625 -1902.8349609375 L -327.1530151367188 -1902.840942382812 L -327.1960144042969 -1902.983032226562 L -326.7680053710938 -1903.201049804688 Z M -339.8739929199219 -1904.406982421875 L -339.8739929199219 -1904.426025390625 L -339.656005859375 -1904.426025390625 L -339.6369934082031 -1904.5 L -339.281005859375 -1904.5 L -339.281005859375 -1904.344970703125 L -339.3840026855469 -1904.2080078125 L -339.8739929199219 -1904.2080078125 L -339.8739929199219 -1904.406982421875 L -339.8739929199219 -1904.406982421875 Z M -339.5260009765625 -1903.921997070312 C -339.5260009765625 -1903.921997070312 -339.3080139160156 -1903.958984375 -339.2900085449219 -1903.958984375 C -339.27099609375 -1903.958984375 -339.2900085449219 -1903.741943359375 -339.2900085449219 -1903.741943359375 L -339.7799987792969 -1903.711059570312 L -339.8739929199219 -1903.822998046875 L -339.5260009765625 -1903.921997070312 Z M -324.5050048828125 -1899.609008789062 L -325.2999877929688 -1899.609008789062 L -325.7850036621094 -1899.969970703125 L -326.2950134277344 -1899.921020507812 L -326.2950134277344 -1899.609008789062 L -326.4559936523438 -1899.609008789062 L -326.6300048828125 -1899.734008789062 L -327.5130004882812 -1899.9580078125 L -327.5130004882812 -1900.530029296875 L -328.6319885253906 -1900.442016601562 L -328.97900390625 -1900.255981445312 L -329.4230041503906 -1900.255981445312 L -329.6409912109375 -1900.277954101562 L -330.1799926757812 -1899.97900390625 L -330.1799926757812 -1899.4150390625 L -331.2829895019531 -1898.619995117188 L -331.1910095214844 -1898.280029296875 L -330.9679870605469 -1898.280029296875 L -331.0260009765625 -1897.9560546875 L -331.1839904785156 -1897.899047851562 L -331.1919860839844 -1897.052978515625 L -330.239013671875 -1895.968017578125 L -329.8250122070312 -1895.968017578125 L -329.7999877929688 -1896.034057617188 L -329.0539855957031 -1896.034057617188 L -328.8389892578125 -1896.233032226562 L -328.4169921875 -1896.233032226562 L -328.1849975585938 -1896.0009765625 L -327.5539855957031 -1895.93603515625 L -327.6380004882812 -1895.098022460938 L -326.9389953613281 -1893.863037109375 L -327.3070068359375 -1893.159057617188 L -327.2820129394531 -1892.828002929688 L -326.9920043945312 -1892.537963867188 L -326.9920043945312 -1891.740966796875 L -326.6109924316406 -1891.22900390625 L -326.6109924316406 -1890.567016601562 L -326.2820129394531 -1890.567016601562 C -328.1130065917969 -1888.318969726562 -330.8989868164062 -1886.880004882812 -334.0159912109375 -1886.880004882812 C -339.5130004882812 -1886.879028320312 -343.9859924316406 -1891.35302734375 -343.9859924316406 -1896.850952148438 C -343.9859924316406 -1898.234985351562 -343.7019958496094 -1899.553955078125 -343.1900024414062 -1900.753051757812 L -343.1900024414062 -1901.06396484375 L -342.8340148925781 -1901.496948242188 C -342.7099914550781 -1901.73095703125 -342.5769958496094 -1901.958984375 -342.4349975585938 -1902.181030273438 L -342.4190063476562 -1902 L -342.8330078125 -1901.496948242188 C -342.9609985351562 -1901.255004882812 -343.0809936523438 -1901.006958007812 -343.1900024414062 -1900.753051757812 L -343.1900024414062 -1900.18603515625 L -342.7749938964844 -1899.985961914062 L -342.7749938964844 -1899.197021484375 L -342.3779907226562 -1898.52001953125 L -342.0549926757812 -1898.469970703125 L -342.0130004882812 -1898.702026367188 L -342.3949890136719 -1899.2900390625 L -342.468994140625 -1899.862060546875 L -342.2460021972656 -1899.862060546875 L -342.1520080566406 -1899.274047851562 L -341.6010131835938 -1898.469970703125 L -341.7430114746094 -1898.2099609375 L -341.3919982910156 -1897.673950195312 L -340.5220031738281 -1897.458984375 L -340.5220031738281 -1897.599975585938 L -340.1740112304688 -1897.550048828125 L -340.2070007324219 -1897.302001953125 L -339.9339904785156 -1897.251953125 L -339.510986328125 -1897.135986328125 L -338.9150085449219 -1896.45703125 L -338.1530151367188 -1896.399047851562 L -338.0780029296875 -1895.776977539062 L -338.6000061035156 -1895.412963867188 L -338.6239929199219 -1894.858032226562 L -338.6990051269531 -1894.517944335938 L -337.9450073242188 -1893.573974609375 L -337.8869934082031 -1893.25 C -337.8869934082031 -1893.25 -337.6130065917969 -1893.176025390625 -337.5799865722656 -1893.176025390625 C -337.5469970703125 -1893.176025390625 -336.9670104980469 -1892.737060546875 -336.9670104980469 -1892.737060546875 L -336.9670104980469 -1891.030029296875 L -336.760009765625 -1890.972045898438 L -336.9010009765625 -1890.18505859375 L -336.5530090332031 -1889.719970703125 L -336.6170043945312 -1888.93994140625 L -336.1570129394531 -1888.131958007812 L -335.5669860839844 -1887.615966796875 L -334.9729919433594 -1887.60498046875 L -334.9150085449219 -1887.796020507812 L -335.3510131835938 -1888.162963867188 L -335.3259887695312 -1888.345947265625 L -335.2479858398438 -1888.568969726562 L -335.2319946289062 -1888.796997070312 L -335.5260009765625 -1888.806030273438 L -335.6749877929688 -1888.993041992188 L -335.4299926757812 -1889.22802734375 L -335.3970031738281 -1889.406005859375 L -335.6709899902344 -1889.484008789062 L -335.6549987792969 -1889.650024414062 L -335.2650146484375 -1889.708984375 L -334.6719970703125 -1889.993041992188 L -334.4739990234375 -1890.358032226562 L -333.8519897460938 -1891.154052734375 L -333.9930114746094 -1891.776000976562 L -333.802001953125 -1892.107055664062 L -333.2309875488281 -1892.088989257812 L -332.8460083007812 -1892.39404296875 L -332.7219848632812 -1893.593994140625 L -332.2950134277344 -1894.135986328125 L -332.2200012207031 -1894.484985351562 L -332.6099853515625 -1894.609008789062 L -332.8670043945312 -1895.031005859375 L -333.7449951171875 -1895.0400390625 L -334.4410095214844 -1895.305053710938 L -334.4739990234375 -1895.802001953125 L -334.7059936523438 -1896.2080078125 L -335.3359985351562 -1896.217041015625 L -335.7000122070312 -1896.787963867188 L -336.0230102539062 -1896.946044921875 L -336.0400085449219 -1896.77197265625 L -336.6279907226562 -1896.737060546875 L -336.8429870605469 -1897.036987304688 L -337.4570007324219 -1897.161010742188 L -337.9620056152344 -1896.577026367188 L -338.7569885253906 -1896.712036132812 L -338.8150024414062 -1897.609008789062 L -339.39599609375 -1897.7080078125 L -339.1629943847656 -1898.14794921875 L -339.22900390625 -1898.400024414062 L -339.9920043945312 -1897.890014648438 L -340.4719848632812 -1897.948974609375 L -340.6430053710938 -1898.323974609375 L -340.5390014648438 -1898.7099609375 L -340.2739868164062 -1899.197998046875 L -339.6650085449219 -1899.505981445312 L -338.489013671875 -1899.506958007812 L -338.4920043945312 -1899.14794921875 L -338.0690002441406 -1898.951049804688 L -338.1029968261719 -1899.56396484375 L -337.7980041503906 -1899.870971679688 L -337.1830139160156 -1900.275024414062 L -337.1409912109375 -1900.558959960938 L -336.5280151367188 -1901.197021484375 L -335.8760070800781 -1901.557983398438 L -335.9339904785156 -1901.60595703125 L -335.4920043945312 -1902.02099609375 L -335.3299865722656 -1901.97802734375 L -335.2560119628906 -1901.885009765625 L -335.0889892578125 -1902.071044921875 L -335.0469970703125 -1902.089965820312 L -335.2309875488281 -1902.115966796875 L -335.4169921875 -1902.177978515625 L -335.4169921875 -1902.35595703125 L -335.3190002441406 -1902.43701171875 L -335.1010131835938 -1902.43701171875 L -335.0010070800781 -1902.39404296875 L -334.9150085449219 -1902.218994140625 L -334.8099975585938 -1902.234985351562 L -334.8099975585938 -1902.2509765625 L -334.7789916992188 -1902.239990234375 L -334.4739990234375 -1902.286987304688 L -334.4299926757812 -1902.43603515625 L -334.2560119628906 -1902.392944335938 L -334.2560119628906 -1902.23095703125 L -334.4179992675781 -1902.119995117188 L -334.4169921875 -1902.119995117188 L -334.3940124511719 -1901.942016601562 L -333.8399963378906 -1901.77197265625 L -333.8389892578125 -1901.765014648438 L -333.7109985351562 -1901.776000976562 L -333.7030029296875 -1902.015991210938 L -334.1419982910156 -1902.214965820312 L -334.1669921875 -1902.3310546875 L -333.802001953125 -1902.454956054688 L -333.7860107421875 -1902.802978515625 L -334.1669921875 -1903.035034179688 L -334.1919860839844 -1903.623046875 L -334.7139892578125 -1903.365966796875 L -334.9039916992188 -1903.365966796875 L -334.85400390625 -1903.81396484375 L -335.5650024414062 -1903.98095703125 L -335.8590087890625 -1903.759033203125 L -335.8590087890625 -1903.08203125 L -336.3880004882812 -1902.913940429688 L -336.5989990234375 -1902.473022460938 L -336.8290100097656 -1902.43603515625 L -336.8290100097656 -1903 L -337.3259887695312 -1903.068969726562 L -337.5750122070312 -1903.23095703125 L -337.6749877929688 -1903.597045898438 L -336.7860107421875 -1904.116943359375 L -336.3510131835938 -1904.25 L -336.3070068359375 -1903.95703125 L -336.0650024414062 -1903.969970703125 L -336.0459899902344 -1904.116943359375 L -335.7929992675781 -1904.152954101562 L -335.7879943847656 -1904.203979492188 L -335.8970031738281 -1904.25 L -335.9219970703125 -1904.405029296875 L -335.6109924316406 -1904.431030273438 L -335.4230041503906 -1904.626953125 L -335.4129943847656 -1904.640991210938 L -335.4100036621094 -1904.640991210938 L -335.3529968261719 -1904.699951171875 L -334.697998046875 -1904.781982421875 L -334.4089965820312 -1904.536010742188 L -335.1679992675781 -1904.131958007812 L -334.2019958496094 -1903.904052734375 L -334.0769958496094 -1904.22705078125 L -333.6539916992188 -1904.22705078125 L -333.5050048828125 -1904.509033203125 L -333.8039855957031 -1904.5830078125 L -333.8039855957031 -1904.93896484375 L -334.739990234375 -1905.35400390625 L -335.385986328125 -1905.280029296875 L -335.7510070800781 -1905.088989257812 L -335.7260131835938 -1904.625 L -336.1069946289062 -1904.682983398438 L -336.1659851074219 -1904.93896484375 L -335.7999877929688 -1905.27099609375 L -336.4630126953125 -1905.303955078125 L -336.6539916992188 -1905.246948242188 L -336.7359924316406 -1905.02294921875 L -336.4880065917969 -1904.98095703125 L -336.5369873046875 -1904.733032226562 L -336.9599914550781 -1904.7080078125 L -337.0260009765625 -1904.5419921875 L -337.6380004882812 -1904.525024414062 C -337.6380004882812 -1904.525024414062 -337.6549987792969 -1904.873046875 -337.6799926757812 -1904.873046875 C -337.7049865722656 -1904.873046875 -337.2000122070312 -1904.880981445312 -337.2000122070312 -1904.880981445312 L -336.8349914550781 -1905.238037109375 L -337.0339965820312 -1905.338012695312 L -337.2990112304688 -1905.079956054688 L -337.739013671875 -1905.10498046875 L -338.0039978027344 -1905.469970703125 L -338.5669860839844 -1905.469970703125 L -339.1549987792969 -1905.031005859375 L -338.6159973144531 -1905.031005859375 L -338.5669860839844 -1904.873046875 L -338.7070007324219 -1904.743041992188 L -338.1099853515625 -1904.725952148438 L -338.0190124511719 -1904.510986328125 L -338.6910095214844 -1904.536010742188 L -338.7229919433594 -1904.702026367188 L -339.14599609375 -1904.79296875 L -339.3699951171875 -1904.9169921875 L -339.8720092773438 -1904.912963867188 C -338.2269897460938 -1906.112060546875 -336.2030029296875 -1906.822021484375 -334.0150146484375 -1906.822021484375 C -331.4939880371094 -1906.822021484375 -329.1900024414062 -1905.879028320312 -327.4320068359375 -1904.3310546875 L -327.5509948730469 -1904.119018554688 L -328.010986328125 -1903.93896484375 L -328.2049865722656 -1903.72802734375 L -328.1600036621094 -1903.483032226562 L -327.9230041503906 -1903.449951171875 L -327.7799987792969 -1903.093017578125 L -327.3699951171875 -1903.256958007812 L -327.3030090332031 -1902.779052734375 L -327.4259948730469 -1902.779052734375 L -327.7619934082031 -1902.828979492188 L -328.135009765625 -1902.765991210938 L -328.4960021972656 -1902.256958007812 L -329.010986328125 -1902.176025390625 L -329.0859985351562 -1901.734008789062 L -328.8680114746094 -1901.682983398438 L -328.9309997558594 -1901.399047851562 L -329.4429931640625 -1901.501953125 L -329.9129943847656 -1901.399047851562 L -330.0130004882812 -1901.137939453125 L -329.9309997558594 -1900.588989257812 L -329.656005859375 -1900.4599609375 L -329.1929931640625 -1900.463012695312 L -328.8810119628906 -1900.490966796875 L -328.7860107421875 -1900.739013671875 L -328.2980041503906 -1901.373046875 L -327.9760131835938 -1901.307983398438 L -327.6600036621094 -1901.593994140625 L -327.6010131835938 -1901.369995117188 L -326.8219909667969 -1900.844970703125 L -326.9179992675781 -1900.717041015625 L -327.2680053710938 -1900.735961914062 L -327.1340026855469 -1900.545043945312 L -326.9179992675781 -1900.496948242188 L -326.6650085449219 -1900.60302734375 L -326.6700134277344 -1900.907958984375 L -326.5580139160156 -1900.963989257812 L -326.6480102539062 -1901.06005859375 L -327.1669921875 -1901.348999023438 L -327.3039855957031 -1901.734008789062 L -326.8729858398438 -1901.734008789062 L -326.7340087890625 -1901.597045898438 L -326.3630065917969 -1901.276977539062 L -326.3469848632812 -1900.889038085938 L -325.9620056152344 -1900.47900390625 L -325.8179931640625 -1901.041015625 L -325.552001953125 -1901.18701171875 L -325.5010070800781 -1900.72705078125 L -325.2409973144531 -1900.441040039062 L -324.7210083007812 -1900.449951171875 C -324.6210021972656 -1900.191040039062 -324.5299987792969 -1899.927978515625 -324.4509887695312 -1899.660034179688 L -324.5050048828125 -1899.609008789062 Z M -338.9540100097656 -1902.779052734375 L -338.6929931640625 -1902.902954101562 L -338.4570007324219 -1902.847045898438 L -338.5379943847656 -1902.530029296875 L -338.7929992675781 -1902.448974609375 L -338.9540100097656 -1902.779052734375 Z M -337.5690002441406 -1902.032958984375 L -337.5690002441406 -1901.828002929688 L -338.1650085449219 -1901.828002929688 L -338.3880004882812 -1901.890014648438 L -338.3320007324219 -1902.032958984375 L -338.0459899902344 -1902.151977539062 L -337.6549987792969 -1902.151977539062 L -337.6549987792969 -1902.032958984375 L -337.5690002441406 -1902.032958984375 Z M -337.2940063476562 -1901.746948242188 L -337.2940063476562 -1901.548950195312 L -337.4440002441406 -1901.452026367188 L -337.6300048828125 -1901.41796875 L -337.6300048828125 -1901.746948242188 L -337.2940063476562 -1901.746948242188 Z M -337.4620056152344 -1901.828002929688 L -337.4620056152344 -1902.06494140625 L -337.2569885253906 -1901.878051757812 L -337.4620056152344 -1901.828002929688 Z M -337.3689880371094 -1901.349975585938 L -337.3689880371094 -1901.156982421875 L -337.5119934082031 -1901.014038085938 L -337.8280029296875 -1901.014038085938 L -337.7789916992188 -1901.23095703125 L -337.6289978027344 -1901.244018554688 L -337.5979919433594 -1901.317993164062 L -337.3689880371094 -1901.349975585938 Z M -338.1579895019531 -1901.746948242188 L -337.8290100097656 -1901.746948242188 L -338.2510070800781 -1901.156982421875 L -338.4249877929688 -1901.25 L -338.3880004882812 -1901.4990234375 L -338.1579895019531 -1901.746948242188 Z M -336.8089904785156 -1901.41796875 L -336.8089904785156 -1901.224975585938 L -337.1260070800781 -1901.224975585938 L -337.2130126953125 -1901.349975585938 L -337.2130126953125 -1901.530029296875 L -337.18798828125 -1901.530029296875 L -336.8089904785156 -1901.41796875 Z M -337.1019897460938 -1901.682983398438 L -337.0119934082031 -1901.777954101562 L -336.8599853515625 -1901.682983398438 L -336.9819946289062 -1901.5830078125 L -337.1019897460938 -1901.682983398438 Z M -324.343994140625 -1899.10595703125 L -324.31201171875 -1899.142944335938 C -324.2990112304688 -1899.087036132812 -324.2860107421875 -1899.030029296875 -324.2730102539062 -1898.973022460938 L -324.343994140625 -1899.10595703125 Z"></path></svg>'
